+++ test/harness/core/types.ts
@@ -92,3 +92,39 @@ export interface PeerHarnessState<
     activeForkId?: ForkId;
     options: Required<HarnessOptions<TFactories>>;
 }
+
+/**
+ * Test context fields used by blocks for cross-block state sharing
+ * These fields are set by certain blocks and consumed by others
+ */
+export interface HarnessContext {
+    /** Original fork ID captured before dispute/fork change (set by Event.captureOriginalFork) */
+    originalForkId?: ForkId;
+
+    /** New fork ID after fork change (set by Context.updateActiveFork) */
+    newForkId?: ForkId;
+
+    /** Index of the malicious peer in Byzantine attack scenarios (set by Context.markMaliciousPeer, Byzantine blocks) */
+    maliciousPeerIndex?: number;
+
+    /** Last malicious peer index from most recent Byzantine attack (set by Byzantine blocks) */
+    lastMaliciousPeerIndex?: number;
+
+    /** Indices of honest peers in Byzantine attack scenarios (set by Context.markMaliciousPeer) */
+    honestPeerIndices?: number[];
+
+    /** Last tampered dispute object (set by Byzantine blocks) */
+    lastTamperedDispute?: any;
+
+    /** Promise that resolves to tampered dispute (set by Byzantine.interceptDisputeConstruction) */
+    tamperedDisputePromise?: Promise<any>;
+
+    /** Function to restore dispute construction after interception (set by Byzantine.interceptDisputeConstruction) */
+    restoreDisputeConstruction?: () => void;
+
+    /** Dynamic snapshot count storage for named contexts - indexed by context key (set by Assert.storeSnapshotCount) */
+    [key: `snapshotCount_${string}`]: number;
+
+    /** Dynamic snapshot count storage for peers - indexed by peer index (set by Context blocks) */
+    [key: `peer${number}SnapshotCountBefore`]: number;
+}
